Refactor backend routes into modules
