+++ backend/llm.py
@@ -1,0 +1,9 @@
+import os
+from dotenv import load_dotenv
+import google.generativeai as genai
+
+load_dotenv()
+
+genai.configure(api_key=os.getenv("GEMINI_API_KEY"))
+
+gemini_model = genai.GenerativeModel(model_name="models/gemini-2.5-flash")
